chore: skip chunk if embedding failed
